# Make `step_textfeatures()` faster