--- a/NEWS.md
+++ b/NEWS.md
@@ -4,6 +4,8 @@
 
 * Calling `?tidy.step_*()` now sends you to the documentation for `step_*()` where the outcome is documented. (#261)
 
+* `step_textfeatures()` has been made faster and more robust. #265
+
 # textrecipes 1.0.6
 
 * textfeatures has been removed from Suggests. (#255)