--- a/R/count_functions.R
+++ b/R/count_functions.R
@@ -42,22 +42,12 @@ n_uq_charS <- function(x) {
 
 
 n_digits <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("\\d", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_regex(x, "\\d")
 }
 
 
 n_hashtags <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("#[[:alnum:]_]+", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_regex(x, "#[[:alnum:]_]+")
 }
 
 n_uq_hashtags <- function(x) {
@@ -72,12 +62,7 @@ n_uq_hashtags <- function(x) {
 }
 
 n_mentions <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("@\\S+", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_regex(x, "@\\S+")
 }
 
 n_uq_mentions <- function(x) {
@@ -92,66 +77,31 @@ n_uq_mentions <- function(x) {
 }
 
 n_commas <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr(",", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_fixed(x, ",")
 }
 
 n_periods <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("\\.", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_fixed(x, ".")
 }
 
 n_exclaims <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("\\!", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_fixed(x, "!")
 }
 
 n_extraspaces <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("\\s{2}|\\t|\\n", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_regex(x, "\\s{2}|\\t|\\n")
 }
 
 n_caps <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("[[:upper:]]", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_regex(x, "[[:upper:]]")
 }
 
 n_lowers <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("[[:lower:]]", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_regex(x, "[[:lower:]]")
 }
 
 n_urls <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("https?", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_regex(x, "https?")
 }
 
 n_uq_urls <- function(x) {
@@ -166,23 +116,13 @@ n_uq_urls <- function(x) {
 }
 
 n_nonasciis <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
   x <- iconv(x, from = "UTF-8", to = "ASCII", sub = "[NONASCII]")
-  m <- gregexpr("\\[NONASCII\\]", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_regex(x, "\\[NONASCII\\]")
 }
 
 n_puncts <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
   x <- gsub("!|\\.|\\,", "", x)
-  m <- gregexpr("[[:punct:]]", x)
-  x <- purrr::map_int(m, ~ sum(.x > 0, na.rm = TRUE))
-  x[na] <- NA_integer_
-  x
+  stringi::stri_count_regex(x, "[[:punct:]]")
 }
 
 first_person <- function(x) {