--- a/R/count_functions.R
+++ b/R/count_functions.R
@@ -8,25 +8,18 @@ n_uq_words <- function(x) {
 }
 
 n_charS <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  x <- gsub("\\s", "", x)
-  x <- nchar(x)
-  x[na] <- NA_integer_
-  x
+  x <- stringi::stri_replace_all_regex(x, "\\s", "")
+  nchar(x)
 }
 
 n_uq_charS <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  x <- gsub("\\s", "", x)
-  x <- strsplit(x, "")
-  x <- lapply(x, unique)
-  x <- lengths(x)
-  x[na] <- NA_integer_
-  x
+  x <- stringi::stri_replace_all_regex(x, "\\s", "")
+  x <- stringi::stri_split_boundaries(
+    x,
+    opts_brkiter = stringi::stri_opts_brkiter(type = "character")
+  )
+  purrr::map_int(x, dplyr::n_distinct)
 }
-
 
 n_digits <- function(x) {
   stringi::stri_count_regex(x, "\\d")
@@ -38,14 +31,8 @@ n_hashtags <- function(x) {
 }
 
 n_uq_hashtags <- function(x) {
-  na <- is.na(x)
-  if (all(na)) return(0)
-  m <- gregexpr("#[[:alnum:]_]+", x)
-  x <- regmatches(x, m)
-  x <- lapply(x, unique)
-  x <- lengths(x)
-  x[na] <- NA_integer_
-  x
+  x <- stringi::stri_extract_all_regex(x, "#[[:alnum:]_]+", omit_no_match = TRUE)
+  purrr::map_int(x, dplyr::n_distinct)
 }
 
 n_mentions <- function(x) {