I found that a big part of the slowness of https://github.com/tidymodels/tidymodels.org/issues/27 came from `step_textfeatures()`, this PR aims to deal with that issue


```r
library(tidymodels)
library(textrecipes)

data("small_fine_foods")

rec <- recipe(~review, data = training_data) |>
  step_textfeature(review)

bench::mark(
  prep(rec)
)

# Before
# A tibble: 1 × 13
# expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
# <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
# 1 prep(rec)     1.59s    1.59s     0.628     679MB     6.28     1    10
# ℹ 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#   time <list>, gc <list>

# Now
# A tibble: 1 × 13
# expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
# <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
# 1 prep(rec)     630ms    630ms      1.59    48.2MB     7.94     1     5
# ℹ 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#   time <list>, gc <list>
```